Removing x_name and df from probability_plot interface
Arguments x_name and df have been removed from the interface of probability_plot function and argument series has been used instead. The purpose of this change is to make the function more flexible. It is now able to handle pandas series, numpy arrays as well as normal python lists.