Fixing a bug causing the pdf to be improperly scaled
In the previous version of the sub-module histogram, the pdf plot and the histogram shared the same y-axis. This caused the pdf to be improperly scaled and become flat when the number of observations is large. To address this issue, a new Axes object has been added which has a separate y-axis to accommodate the pdf plot.